chore: take care of edge case when more than 16 boards are in use

+++ main/helpers.h
@@ -4,8 +4,8 @@
 #include "main.h"
 #include <Arduino.h>
 
-int getBoardIndexForChannel(int channel);
-int getBoardAddressForChannel(int channel);
-int getBoardSubAddressForChannel(int channel);
+uint8_t getBoardIndexForChannel(uint16_t channel);
+uint8_t getBoardAddressForChannel(uint16_t channel);
+uint8_t getBoardSubAddressForChannel(uint16_t channel);
 
 #endif // helpers_h
